Notebook
Early stage diabetes, also known as prediabetes, occurs when blood sugar levels are higher than normal but not high enough to be classified as diabetes. In this stage, the body becomes insulin resistant, making it difficult for glucose to enter cells, leading to high glucose levels in the bloodstream. Common symptoms include increased thirst and urination, fatigue, and blurred vision. Untreated, early stage diabetes can progress to type 2 diabetes or even type 1 diabetes. Lifestyle changes, such as a healthy diet and regular exercise, can often reverse or prevent the onset of early stage diabetes.  

This model helps in predicting the diabetes based on the values entered by the user with an enhanced accuracy.